soccer-stats/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0